requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
phonenumbers==8.13.53
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whatsapp-api-client-python==0.0.47
